Changed CMakeLists.txt code
Now that we are using a custom cross compiler, we will no longer need to use --target (at least, we see no need right now) as a c/c++ flag. Clang and clang++ are still the default cross compilers, so it is recommended you change the default compilers in src/CMakeLists.txt if you are using what we have compiled. This will not be optional in the future, as building for i686-pc-none-elf will be automatically done when compiling this project.